Disassociate registry and loader management
