--- specviz/io/wizard/__init__.py
@@ -1,1 +1,0 @@
-
